perf: fix comment
